Move RmiLogger impl to cpp file
